Fix crash when entering Save Select screen
GameMaker Studio version 2.2.2 changed the behavior of the function
'real'; previously, it would silently convert invalid strings to '0',
but now those strings now cause the function to throw an error and
crash the game. This behavior was relied upon when loading level data
into a save slot, so after importing to GMS2, the game would crash
when entering the Save Select screen. To fix this, a new function
'__real' was created that implements the original behavior of 'real'.

Credit shall be given to YellowAfterLife before the release of 1.0.0.