Replace 'layer' with 'depth_mask'
The variable 'layer' is a bitmask used to simulate the third
dimension. Each bit represents a level of depth that can be
interacted with. The player character cannot interact with certain
objects unless their 'layer' matches bitwise with the objects'
'layer'.

In GameMaker Studio 2, 'layer' became a built-in variable used for
managing the draw priority of objects in a room. This variable is a
pointer reference, rather than a binary number. GMS2 does not
automatically convert previous references to this variable after
importing from GameMaker: Studio; as a result, several objects were
now having their depth layers changed to an invalid index. This
resulted in a crash when attempting to enter levels.

All previous references to 'layer' have been replaced with
'depth_mask' instead to avoid the name clash.

Fixes #4