Add a Contribution Guide
These are the standards which all contributors are expected to uphold.
They will appear whenever someone attempts to submit an issue or pull
request. There will also be a link to the contribution guide in the
readme. This will give everyone enough chances to learn the rules
before consequences are administered.