Fix soft-lock when moving on certain terrain
The script `player_get_terrain` is used to confirm the terrain the
player is standing on. It will attempt to align the player's position
on top of the terrain, first by moving upward if the player is stuck
inside, then moving downward if the terrain is actually further below
the player's 'feet'. These checks use while loops, as there's a certain
confidence that the terrain previously found can be found again using
these algorithms. Inexplicably, some terrain would be found closer to
the player's center than below their feet, but would not get caught by
the check for moving upwards when inside the terrain. Since that
places the next check past the point where the terrain could even be
detected at all, it resulted in the respective while loop continuing
indefinitely, creating a soft-lock.

No such soft-locks occur in the original release of the game, so there
is still enough confidence in the acquired terrain to rely on while
loops for the algorithm. Instead, the algorithm was slightly adjusted
to scan the area from the player's centre point to their 'feet'
before attempting to move downward to find the terrain, stopping if
the terrain is found earlier than expected. This covers the edge
cases that were causing the soft-lock.

Fixes #8