Extend platform at Tidal Tubes Act 2 capsule
It was possible for the player character to move fast enough after
breaking open the end-of-level capsule in Tidal Tubes Zone Act 2
(Future or Past) that they go off the far-right edge of the platform;
in doing so, they'd become stuck when attempting to move automatically
back to the center of the screen, causing a soft-lock. To prevent
this, the platform has been extended further to the right in both
versions of the level. Furthermore, adjustments were made to the
level tiles to help keep the terrain looking cohesive.

Fixes #30