Resolve "Solid" object variable references
The script PlayerCollision references variables which are not properly
initialized by the object "Solid," but were set in the script SolidNew,
which was removed from the project. Even though SolidNew was unused,
the YoYo Compiler (YYC) still considered it a valid route to
initializing those variables; when the script was removed, it treated
those variables as uninitialized and refused to compile.

The variables are now properly initialized in the Create event of
"Solid", resolving the issue. Note that it is likely that "Solid"
itself is an unused asset and that these changes will be rendered
obsolete in a future commit.