Remove notes and compatibility report
The default notes file is empty, as the 'Game Information' file was not
made use of in the original GameMaker: Studio project. It has been
removed, but may be regenerated later should the need arise.

The compatibility report lists all changes made to the project after
importing it into GameMaker Studio 2. Github now contains a copy of
these changes, so it is unnecessary to maintain it moving forward and
it can be safely deleted.