Sustain AAZ2 boundary changes after time travel
After defeating the mini-boss at the start of Attraction Attack Zone
Act 2 (either Past or Future,) the left camera boundary is moved ahead
to prevent the player character from returning to the mini-boss' arena;
then, there is a short intermission area before the 'actual' level.
Reaching the end of this moves the camera boundary further forward to
prevent backtracking into the intermission area. However, the player
could still enter after time traveling to another version of the Act
(or restarting from a checkpoint.)

The various starting positions in both versions of the act now set
the proper left camera boundary when the playable character is
spawned in, so it should no longer be possible to return to the
intermission area once the player has passed it.