Fix ring count bug during Galacnik Gauntlet
It is no longer possible for the player's ring count to fall below 0
after taking a hit in Galacnik Gauntlet Zone. Instead of waiting for
the playable character's superform to naturally expire, taking damage
that would reduce the ring count to 0 or less will immediately
de-transform the playable character.

Fixes #55.