Improve Sealbot enemy behavior
The behavior of the Sealbot enemy was causing performance issues on
lower-end devices. This presented an opportunity to refactor the object
code and make adjustments to enemy's design.

- The enemy no longer activates unless it is on-screen. This can
potentially result in the enemy getting stuck in newly-activated
terrain, but it should improve performance.

- The enemy now animates differently, particularly when charging at a
playable character.

- The enemy will continue its pursuit of any playable character in
range until they step out of range.

See also: #31