Establish 'sprite explosion' api
There are many times in the game where a single sprite is 'blown up'
into several pieces that fly out the screen. Currently, these pieces
use a copy of the base sprite that has been split into multiple
sub-images of equal size. This needlessly consumes texture space with
redundant images and slows down compilation.

The script `sprite_explosion_create` takes a base image and creates
'debris' (objSpriteDebris) that can render the image directly, but in
parts. By using this api, the copies can be removed from the project,
saving disk space and achieving a small gain in compilation speed. The
api also grants flexibility in the shape and volume of the 'debris'
that would have been more difficult to achieve previously.

Related: #36