Improve Ostrich enemy behavior
The Ostrich enemy was not behaving correctly when frozen by the Ice
Shield's ice attack. This presented an opportunity to refactor the
object code and make adjustments to the enemy's design.

- The enemy now properly stops moving and animating while frozen; this
also includes the head when it's extended from the ground.

- The enemy now uses sprite_explosion_create to generate debris when
breaking out of an ice block. This behavior now also occurs when it's
destroyed by the playable character while frozen.

- When the enemy is frozen, the ice block is displayed closer to the
enemy's body.

- Destroying the enemy's head after it's extended no longer spawns an
extra 'flower spark'.

- Destroying the enemy's body after the head's extended also destroys
the head.

- The 'flower spark' now spawns closer to the enemy's body after the
enemy is destroyed.