Improve Whalebot enemy behavior
When breaking out of an ice block, the Whalebot enemy would remain
stationary instead of continuing to move from where it was frozen. In
correcting this issue, the opportunity was taken to refactor the object
code and make adjustments to the enemy's design.

- The enemy's normal behavior is restored after breaking out of an
ice block.

- The enemy now properly stops animating while frozen; the past version
of the enemy will also stop playing the propeller sound effect
immediately (assuming no other instances of the enemy are on-screen
and un-frozen.)

- The enemy now uses sprite_explosion_create to generate debris when
breaking out of an ice block. This behavior now also occurs when it's
destroyed by the playable character while frozen.

- The future version of the enemy now uses a separate animation that
is more consistent with its sprite frames; redundant frames were
then cut from the sprite.

- The past version of the enemy stops playing the propeller sound
effect immediately when destroyed by a playable character.

Fixes: #62