Improve Snailbot enemy behavior
The Snailbot enemy was not behaving correctly when frozen by the Ice
Shield's ice attack. This presented an opportunity to refactor the
object code and make adjustments to the enemy's design.

- The enemy now properly stops moving and animating while frozen.

- The enemy now uses sprite_explosion_create to generate debris when
breaking out of an ice block. This behavior now also occurs when it's
destroyed by the playable character while frozen.

Fixes: #63