Improve Dinobot enemy behavior
The Dinobot enemy was not properly designed to react to getting frozen
by the Ice Shield's special ability. This presented an opportunity to
refactor the object code and make adjustments to the enemy's design.

- The enemy now waits for a playable character to enter into a
pre-defined range before 'pouncing'.

- If the enemy is frozen before 'pouncing', it will properly stop until
the condition expires, after which it will 'pounce' immediately. (It is
difficult to observe this scenario because of how quickly the enemy
'pounces' when in range, but the behavior is implemented nonetheless.)

- If the enemy is frozen after 'pouncing', it will stop in midair and
fall helplessly off-screen until it is removed by the game.

- The enemy now uses sprite_explosion_create to generate debris when
breaking out of an ice block. This behavior now also occurs when it's
destroyed by the playable character while frozen.

Fixes: #63