Improve Bubbles enemy behavior
The Bubbles enemy was not properly designed to react to getting frozen
by the Ice Shield's special ability. This presented an opportunity to
refactor the object code and make adjustments to the enemy's design.

- The enemy now properly stops moving and animating while frozen.

- The enemy will immediately go into the 'raging' state after thawing
from being frozen.

- The enemy now uses sprite_explosion_create to generate debris when
breaking out of an ice block. This behavior now also occurs when it's
destroyed by the playable character while frozen.

- The enemy now properly stops the 'shaking' sound effect immediately
when destroyed by a playable character (assuming no other instances
of the enemy are on-screen and not frozen.)

- The bubbles that orbit the enemy are slightly visible at all times,
becoming more opaque when the enemy enters the 'raging' state.

- The orbiting bubbles rotate at different rates relative to their
size (plus a bit of randomness.)

- The enemy can now bring more than one playable character into its
orbit when in the 'raging' state (to accomodate multiplayer
functionality later.)

- Any playable characters trapped in orbit are released at the same
speed and direction that they were orbiting around the enemy when it
leaves the 'raging' state.

Fixes: #64