Refactor objGalanikFireball
The object objGalanikFireball still contained a missed reference to
objBreakableRockBits, which was deleted in a previous commit. This was
preventing compilation to YoYo Compiler (YYC) output. While removing
the reference, the object was also refactored to reduce coupling and
make it easier to change; additionally, the object now moves slightly
faster to ensure it can successfully reach its target. (One possible
side effect is that the associated boss' attack may become a little
more difficult to avoid.)