Force more spawn points to sustain AAZ2 boundary
It was possible to return to the intermission area of Attaction Attack
Zone Act 2 (either Past or Future) by losing a life OR by visiting any
Special Stage through a Special Stage Ring; that area should be blocked
by the camera once the player has passed it. The updated camera
boundary is now sustained in these cases.

Additionally, the starting position after passing the intermission area
in the Future version of level has been moved to align with the same
location in the Past version so the camera placement when starting the
level respects the updated boundary.

Fixes: #64, #65
Related: #54