Remove angle_distance
The script `angle_distance` has been deprecated by the introduction of
the built-in function `angle_difference`. All references to the former
have been replaced by the latter, and the associated code has been
refactored with various other improvements:

- 80 character line length

- Tabs for indentation instead of spaces

- Only one return statement, if easily achievable

In addition, the script `player_floor_transfer` was updated to use the
same code as the script `player_get_terrain` for adapting the
playable character's height to the current floor.